change date to year,month to ::int

+++ dbt/models/gold/gold_customer_trends.sql
@@ -6,8 +6,8 @@
 select
     customer_sk,
     operator,
-    date_part('year', registration_date) as registration_year,
-    date_part('month', registration_date) as registration_month
+    date_part('year', registration_date)::int as registration_year,
+    date_part('month', registration_date)::int as registration_month
 from {{ ref('silver_mobile_customers') }}
 where registration_date is not null
